+++ extensions/debug/Module.php
@@ -209,8 +209,8 @@ class Module extends \yii\base\Module implements BootstrapInterface
             'log' => ['class' => 'yii\debug\panels\LogPanel'],
             'profiling' => ['class' => 'yii\debug\panels\ProfilingPanel'],
             'db' => ['class' => 'yii\debug\panels\DbPanel'],
-            'mail' => ['class' => 'yii\debug\panels\MailPanel'],
             'assets' => ['class' => 'yii\debug\panels\AssetPanel'],
+            'mail' => ['class' => 'yii\debug\panels\MailPanel'],
         ];
     }
 }