+++ extensions/debug/views/default/panels/assets/summary.php
@@ -1,6 +1,8 @@
 <?php
 /* @var $panel yii\debug\panels\AssetPanel */
+if (!empty($panel->data)):
 ?>
 <div class="yii-debug-toolbar-block">
-    <a href="<?= $panel->getUrl() ?>" title="Asset bundles count">Asset bundles <span class="label label-info"><?= $panel->data['totalBundles'] ?></span></a>
+    <a href="<?= $panel->getUrl() ?>" title="Number of asset bundles loaded">Asset Bundles <span class="label label-info"><?= count($panel->data) ?></span></a>
 </div>
+<?php endif; ?>